requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.6
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.3.post0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
kombu==5.2.4
msgpack==1.0.4
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydotplus==2.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.1
reportlab==3.6.11
secret-key-generator==0.0.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
zope.interface==5.4.0
